requirements.txt
openai-whisper
sounddevice
numpy
torch
scipy
opencv-python
mediapipe
